+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Upcoming release]
 
+- Add Lambda warming for Tilegarden
+
 ## [0.9.0] - 2019-02-14
 
 #### Added